Create categories for adiministrative divisions (Perú)

Function 'gen_admin.div' returns the natural, geographical, or administrative division of Peru that includes the region, province or district provided to the function.

gen_admin.div(x, method = "Traditional")

Arguments

x
x is a list/vector with one of the 25 regions/departments names of Peru (first-level administrative subdivisions)

method
defines the type of natural, geographical, or administrative division that should be returned Traditional: Coastal, Andes Mountains, Amazon Jungle, or Lima & Callao TC: (Lima no incluye Huarua/Cañete)Administratives Macro Regions of the Constitutional Tribunal Cardinal: North, South, Amazon Jungle, or Lima & Callao Traditional Cardinal: Coastal and Andes Mountains are divided into North and South (e.g. South Andes Mountains)

Value

Tabla con las divisiones resultantes del Peru

Examples

df <- data.frame(reg=c("LIMA","CALLAO","CAJAMARCA","AMAZONAS","SAN MARTIN","HUANUCO","PASCO",
"JUNIN","CUSCO", "PUNO","APURIMAC","AYACUCHO","HUANCAVELICA","TUMBES","PIURA","LAMBAYEQUE","LA LIBERTAD",
"ANCASH","ICA","AREQUIPA","TACNA","MOQUEGUA","LORETO","UCAYALI", "MADRE DE DIOS"), stringsAsFactors = FALSE)